+++ experimenter/experimenter/nimbus_ui_new/templates/nimbus_experiments/takeaways_card.html
@@ -93,7 +93,7 @@
       <table class="table table-striped">
         <tbody>
           <tr>
-            <th>QBR learning</th>
+            <th>QBR Learning</th>
             <td>{{ experiment.takeaways_qbr_learning }}</td>
             <th>Statistically significant DAU Gain</th>
             <td>{{ experiment.takeaways_metric_gain }}</td>
