+++ experimenter/experimenter/nimbus_ui_new/views.py
@@ -5,7 +5,6 @@ from django.views.generic.edit import UpdateView
 from django_filters.views import FilterView
 
 from experimenter.experiments.constants import RISK_QUESTIONS
-from experimenter.experiments.forms import QAStatusForm, TakeawaysForm
 from experimenter.experiments.models import NimbusExperiment
 from experimenter.nimbus_ui_new.filtersets import (
     STATUS_FILTERS,
@@ -13,6 +12,7 @@ from experimenter.nimbus_ui_new.filtersets import (
     SortChoices,
     StatusChoices,
 )
+from experimenter.nimbus_ui_new.forms import QAStatusForm, TakeawaysForm
 
 
 class NimbusChangeLogsView(DetailView):
